% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree_of_enrollment.R
\name{degree_of_enrollment}
\alias{degree_of_enrollment}
\title{degree_of_enrollment}
\usage{
degree_of_enrollment(
  df,
  anticipated_column,
  actual_column,
  trial_column,
  round_off = 2
)
}
\arguments{
\item{df}{The dataframe containing enrollment data.}

\item{anticipated_column}{The column name for the anticipated
enrollment values in the dataframe.}

\item{actual_column}{The column name for the actual enrollment
values in the dataframe.}

\item{trial_column}{The column for grouping the data by trial (trial id)}

\item{round_off}{(Optional) The number of decimal places to round
the percentage values. Defaults to 2. Use NA to avoid rounding.}
}
\value{
A list containing:
updated_dataframe: The input dataframe with an added column enrollment_percentage,
showing the calculated enrollment percentage for each row.
trial_summary: A data frame summarizing enrollment data by trial level (total anticipated,
total actual, and trial-level enrollment percentages).
dataset_summary: A data frame summarizing dataset-wide enrollment statistics (median, minimum, maximum)
summary_anticipated_enrollment: A summary of anticipated enrollment statistics (median, IQR, and quantiles).
summary_actual_enrollment: A summary of actual enrollment statistics (median, IQR, and quantiles).
overall_degree_of_enrollment: A summary of the overall enrollment percentage
}
\description{
This function checks the enrollment status in a dataframe by
calculating the enrollment percentage based on the values in the
'anticipated_enrollment' and 'actual_enrollment' columns.  It also computes
trial and dataset summaries and additional statistics related to the
anticipated and actual enrollments.
}
\examples{
result <- degree_of_enrollment(df, "anticipated_enrollment", "actual_enrollment", "trial_id")
print(result$trial_summary)
}
