% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration_of_trial.R
\name{duration_of_trial}
\alias{duration_of_trial}
\title{duration_of_trials}
\usage{
duration_of_trial(df, start_date_col, stop_date_col, inclusive = TRUE)
}
\arguments{
\item{df}{The data frame containing enrolment data.}

\item{start_date_col}{Character string. The column name indicating
start start date of the trial.}

\item{stop_date_col}{Character string. The column name indicating
date on which its corresponding clinical trial record's overall
status was first changed to "Terminated" from any other overall
status in the ClinicalTrials.gov registry.}

\item{inclusive}{Boolean TRUE or FALSE. If TRUE, it adds one to the
number of trial days calculated to include the last day. This
would mean that a trial with a start date on the 1st of the
month and a stop date on the 31st of the month would have a
duration of 31 days (31-1+1, default behaviour). If FALSE,
nothing is added to the number of trial days calculated. This
would mean that a trial with a start date on the 1st of the
month and a stop date on the 31st of the month would have a
duration of 30 days only (31-1+0).}
}
\value{
The original input data frame with an additional column
named 'trial_days' representing the duration of trial in
days. If missing values are encountered in either the start or
stop date columns, the 'trial_days' value for that row will be
set to "Check missing values".
}
\description{
This function calculates the duration of trial in days for each row
of a data frame, based on the provided start and stop date columns.
It handles different date formats and provides a warning if the
format of the dates has changed.
}
\examples{

df <- duration_of_trial(df, "start_date", "stop_date")
}
