% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_actual_enrollment.R
\name{get_actual_enrollment}
\alias{get_actual_enrollment}
\title{get_actual_enrollment}
\usage{
get_actual_enrollment(nctid)
}
\arguments{
\item{nctid}{Trial registration number for ClinicalTrials.gov trial. Can be a single value or a vector.}
}
\value{
actual enrollment
}
\description{
This function retrieves the final recorded enrollment of a clinical trial
as reported on ClinicalTrials.gov.
}
\examples{

get_actual_enrollment("NCT00480077")


}
