% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_anticipated_enrollment.R
\name{get_anticipated_enrollment}
\alias{get_anticipated_enrollment}
\title{get_anticipated_enrollment}
\usage{
get_anticipated_enrollment(nctid)
}
\arguments{
\item{nctid}{Trial registration number for ClinicalTrials.gov trial. Can be a single value or a vector.}
}
\value{
anticipated enrollment
}
\description{
This function retrieves the estimated enrollment of a clinical trial
}
\examples{

get_anticipated_enrollment("NCT00480077")


}
