% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ctgov_sae_group_data.R
\name{get_ctgov_sae_groups}
\alias{get_ctgov_sae_groups}
\title{get_ctgov_sae_groups}
\usage{
get_ctgov_sae_groups(nctid)
}
\arguments{
\item{nctid}{Trial registration number for ClinicalTrials.gov trial. Can be a single value or a vector.}
}
\value{
A tibble with the trial registration number, group ID, title, description, serious number affected, and serious number at risk for each SAE group.
}
\description{
This function retrieves the serious adverse event (SAE) group data from ClinicalTrials.gov for a given clinical trial ID.
}
\examples{
get_ctgov_sae_groups("NCT02453282")


}
