% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_euctr_results.R
\name{get_euctr_results}
\alias{get_euctr_results}
\title{get_euctr_results}
\usage{
get_euctr_results(euctr_ids)
}
\arguments{
\item{euctr_ids}{Trial registration number for EUCTR trial}
}
\value{
Table identifiers list with EUCTR entry and links to their
result section available webpage
}
\description{
Retrieves result link if results were posted for EUCTR
entry, otherwise returns 'No result found'
}
\examples{

get_euctr_results("2007-002223-32")
## result
##'https://www.clinicaltrialsregister.eu/ctr-search/trial/2007-002223-32/results'

get_euctr_results("2005-003854-80")
## no result found ## NA
}
