% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_why_stopped.R
\name{get_why_stopped}
\alias{get_why_stopped}
\title{get_why_stopped}
\usage{
get_why_stopped(nctid, historical_version = FALSE, df = NULL)
}
\arguments{
\item{nctid}{A character string or a vector of character strings
containing a trial registration number for ClinicalTrials.gov.}

\item{historical_version}{A Boolean TRUE or FALSE. If TRUE, the
function expects a data frame containing historical version
data of the type generated by
\code{cthist::clinicaltrials_gov_download()}. If FALSE, this
function will download the historical version data. FALSE by
default.}

\item{df}{A data frame containing historical version data of the
type generated by \code{cthist::clinicaltrials_gov_download()}, will
be used instead of downloading new data if argument
\code{historical_version} is TRUE, and if FALSE, ignored.}
}
\value{
A data frame with two columns: \code{nctid}, which contains
ClinicalTrials.gov trial registration numbers and
\code{reason_for_termination}, which will contain the reason why the
trial was stopped, if provided. If the reason is not available,
\code{reason_for_termination} will contain "\link{No reason posted}".
}
\description{
This function retrieves the reason for trial
termination, if given, for a specified clinical trial.
}
\examples{

get_why_stopped("NCT00480077")
## Low enrollment rates

get_why_stopped(c("NCT04007003", "NCT03693833", "NCT04338971"))
## Sponsor decision, COVID-19 and [No reason posted]
}
