% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_summary_result.R
\name{has_summary_result}
\alias{has_summary_result}
\title{has_summary_result}
\usage{
has_summary_result(nctid)
}
\arguments{
\item{nctid}{Trial registration number for ClinicalTrials.gov trial. Can be a single value or a vector.}
}
\value{
logical vector TRUE if results were posted, otherwise FALSE
}
\description{
This function checks whether a ClinicalTrials.gov trial has posted summary results.
}
\examples{

has_summary_result("NCT00480077") ## TRUE

}
