% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_arms.R
\name{how_many_trial_arms}
\alias{how_many_trial_arms}
\title{how_many_trial_arms}
\usage{
how_many_trial_arms(nctid)
}
\arguments{
\item{nctid}{A string containing a well-formed NCT number for a
trial on ClinicalTrials.gov}
}
\description{
Retrieves the number of trial arms for a trial
registry entry on ClinicalTrials.gov
}
\examples{

how_many_trial_arms("NCT02715531")
## Should return 5

how_many_trial_arms("NCT03430843")
## Should return 2

how_many_trial_arms("NCT03430844")
## Should throw an error (NCT does not exist)

how_many_trial_arms("Lol bad NCT")
## Should throw an error (NCT not well-formed)
}
