% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminatetrialsstudy-package.R
\docType{package}
\name{terminatedtrialsstudy-package}
\alias{terminatedtrialsstudy}
\alias{terminatedtrialsstudy-package}
\title{terminatedtrialsstudy: Processing Terminated Trials from ClinicalTrials.gov}
\description{
Processing Terminated Trials from ClinicalTrials.gov
}
\details{
This package provides functions for analyzing terminated
trials from ClinicalTrials.gov
}
\author{
\strong{Maintainer}: Samruddhi Yerunkar \email{samruddhi.yerunkar@charite.de} (\href{https://orcid.org/0000-0002-0748-383X}{ORCID})

Authors:
\itemize{
  \item Benjamin Gregory Carlisle \email{murph@bgcarlisle.com} (\href{https://orcid.org/0000-0001-8975-0649}{ORCID})
}

}
\keyword{internal}
