% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree_of_enrollment.R
\name{degree_of_enrollment}
\alias{degree_of_enrollment}
\title{degree_of_enrollment}
\usage{
degree_of_enrollment(df, anticipated_column, actual_column, round_off = 2)
}
\arguments{
\item{df}{The dataframe containing enrolment data.}

\item{anticipated_column}{The column name for the anticipated
enrollment values in the dataframe.}

\item{actual_column}{The column name for the actual enrollment
values in the dataframe.}

\item{round_off}{(Optional) The number of decimal places to round
the percentage values. Defaults to 2. Use NA to avoid rounding.}
}
\value{
The updated dataframe with the 'enrollment_percentage'
column populated.
}
\description{
This function checks the enrolment status in a dataframe by
calculating the enrolment percentage based on the values in the
'anticipated_enrolment' and 'actual_enrolment' columns.  It handles
missing values and provides an option to round the percentage
values.

The function iterates through each row of the dataframe and checks
if there are missing values in the 'anticipated_enrollment' and
'actual_enrollment' columns. If any of the values is missing, it
assigns "Check missing values" to the corresponding row in the
'percentage' column. Otherwise, it calculates the enrollment
percentage using the formula: actual_enrollment /
anticipated_enrollment * 100.  The percentage values can be
optionally rounded to a specified number of decimal places.

Usage: degree_of_enrollment(df, "anticipated_column",
"actual_column", round_off = 2)
}
\examples{

  # Assuming you have a dataframe named 'df' with
  # 'anticipated_enrollment' and 'actual_enrollment' columns:

  # Round the percentage values to 1 decimal place
  updated_df <- degree_of_enrollment(
    df,
    "anticipated_enrollment",
    "actual_enrollment",
    round_off = 1
  )

  # Do not round the percentage values
  updated_df <- degree_of_enrollment(
    df,
    "anticipated_enrollment",
    "actual_enrollment",
    round_off = NA
  )
}
