% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_start_date.R
\name{get_start_date}
\alias{get_start_date}
\title{get_start_date}
\usage{
get_start_date(nctid)
}
\arguments{
\item{nctid}{Trial registration number for ClinicalTrials.gov trial. Can be a single value or a vector.}
}
\value{
start date
}
\description{
This function creates the start date of a clinical trial, which is defined
as the study start date when the overall status of the trial was first changed to ‘Terminated’
from any other overall status in the ClinicalTrials.gov registry.
}
\examples{

get_start_date("NCT00480077")

}
