% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stop_date.R
\name{get_stop_date}
\alias{get_stop_date}
\title{get_stop_date}
\usage{
get_stop_date(nctid)
}
\arguments{
\item{nctid}{Trial registration number for ClinicalTrials.gov trial. Can be a single value or a vector.}
}
\value{
stop date
}
\description{
This function creates stop_date of a clinical trial, which is defined
as the actual primary completion date reported on the clinical trial record when
its overall status was first changed to ‘Terminated’ from any other overall status
in the ClinicalTrials.gov registry.
}
\examples{

get_stop_date("NCT00480077")

}
