% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_why_stopped.R
\name{get_why_stopped}
\alias{get_why_stopped}
\title{get_why_stopped}
\usage{
get_why_stopped(nctid)
}
\arguments{
\item{nctid}{Trial registration number for ClinicalTrials.gov trial}
}
\value{
Reason for termination of the clinical trial, or \verb{no result posted}
if the reason is not available.
}
\description{
This function retrieves the reason for trial termination,
if given, for a specified clinical trial.
}
\examples{
get_why_stopped("NCT00480077") ## Low enrollment rates

}
